+++ library/rustsbi/src/lib.rs
@@ -542,6 +542,7 @@
 
 mod console;
 mod cppc;
+mod dbtr;
 mod hsm;
 mod ipi;
 mod mpxy;
@@ -844,6 +845,7 @@ pub use sbi_spec::binary::{CounterMask, HartMask, Physical, SbiRet, SharedPtr};
 /// | `nacl` | [`Nacl`](trait.Nacl.html) | Nested Acceleration extension |
 /// | `sta` | [`Sta`](trait.Sta.html) | Steal Time Accounting extension |
 /// | `mpxy` | [`Mpxy`](trait.Mpxy.html) | Message Proxy extension |
+/// | `dbtr` | [`Dbtr`](trait.Dbtr.html) | Debug Triggers extension |
 ///
 /// The `EnvInfo` parameter is used by RISC-V SBI Base extension which is always supported on all
 /// RISC-V SBI implementations. RustSBI provides the Base extension with additional `EnvInfo` by default.
@@ -1129,6 +1131,7 @@ pub use rustsbi_macros::RustSBI;
 
 pub use console::Console;
 pub use cppc::Cppc;
+pub use dbtr::Dbtr;
 pub use hsm::Hsm;
 pub use ipi::Ipi;
 pub use mpxy::Mpxy;
@@ -1151,12 +1154,13 @@ pub use traits::_rustsbi_base_bare;
 #[doc(hidden)]
 pub use traits::{
     _ExtensionProbe, _StandardExtensionProbe, _rustsbi_base_env_info, _rustsbi_console,
-    _rustsbi_cppc, _rustsbi_fence, _rustsbi_hsm, _rustsbi_ipi, _rustsbi_mpxy, _rustsbi_nacl,
-    _rustsbi_pmu, _rustsbi_reset, _rustsbi_sta, _rustsbi_susp, _rustsbi_timer,
+    _rustsbi_cppc, _rustsbi_dbtr, _rustsbi_fence, _rustsbi_hsm, _rustsbi_ipi, _rustsbi_mpxy,
+    _rustsbi_nacl, _rustsbi_pmu, _rustsbi_reset, _rustsbi_sta, _rustsbi_susp, _rustsbi_timer,
 };
 #[doc(hidden)]
 pub use traits::{
-    _rustsbi_console_probe, _rustsbi_cppc_probe, _rustsbi_fence_probe, _rustsbi_hsm_probe,
-    _rustsbi_ipi_probe, _rustsbi_mpxy_probe, _rustsbi_nacl_probe, _rustsbi_pmu_probe,
-    _rustsbi_reset_probe, _rustsbi_sta_probe, _rustsbi_susp_probe, _rustsbi_timer_probe,
+    _rustsbi_console_probe, _rustsbi_cppc_probe, _rustsbi_dbtr_probe, _rustsbi_fence_probe,
+    _rustsbi_hsm_probe, _rustsbi_ipi_probe, _rustsbi_mpxy_probe, _rustsbi_nacl_probe,
+    _rustsbi_pmu_probe, _rustsbi_reset_probe, _rustsbi_sta_probe, _rustsbi_susp_probe,
+    _rustsbi_timer_probe,
 };