+++ library/macros/CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add RustSBI derive support for the SBI Message Proxy extension.
+- Add RustSBI derive support for the SBI Debug Triggers extension.
 
 ### Modified